requirements.txt
python==3.6.9
requests==2.24.0
python-dotenv==0.14.0
